Adds page list item grid, moves page list item file
