+++ src/scss/components/_pageListItem.scss
@@ -2,10 +2,16 @@
 // Page list item
 // ----------------------------------------
 
+.page-list-items {
+  display: grid;
+  grid-template-columns: 25% 25% 25% 25%;
+  grid-template-rows: 25% 25% 25% 25%;
+}
+
 .page-list-item {
-  height: 300px;
-  width: 200px;
+  min-height: 250px;
   background-size: cover;
+  background-position: center;
   background-repeat: no-repeat;
 }
